Explicitally check for invalid content at a controller level
This adds tests to the sub section controller that tests for invalid content. Previously we had tests for invalid sub sections but this was done by altering the title. Given the changes to how content is managed these tests are required to ensure that it is being handled properly and that error messages are displayed in the expected manner.

Co-authored-by: Alex Newton <alex.newton@digital.cabinet-office.gov.uk>
Co-authored-by: Kevin Dew <kevin.dew@digital.cabinet-office.gov.uk>