Hook the new Structured content up to the JSON presenter
This makes it so that the JSON presenter reads sub section content from the newly created StructuredContent. This also required the removal of a lot of old code and the tests associated with it.

The json presenter has maintained responsibility for creating something that can ultimately be read by the content store. This is achieved by the `sub_sections` method which takes the array of headings and links from the `StructuredContent` model and turns them into something that the content store can work with.

Co-authored-by: Alex Newton <alex.newton@digital.cabinet-office.gov.uk>
Co-authored-by: Kevin Dew <kevin.dew@digital.cabinet-office.gov.uk>