Rename featured_link field on sub_sections to action_link_url
The featured_link field was added to coronavirus_sub_sections here https://github.com/alphagov/collections-publisher/pull/1122. However, it over-loaded the preexisting concept of a featured link (which is automatically generated if the content field contains a topic link).

The rename better describes the intended use of the field, removes the confusion around featured links, and paves the way for the introduction of a more general action_link concept.