Update featured_link to action_link_url
The featured_link field on the coronavirus_sub_sections table has been renamed to action_link_url.

We need to reflect that change where the field is used within the code and permit them in the sub_sections_controller.

We cannot simply change all occurrences of featured_link as the original concept of a featured link is still in use. This concept relates to a link that is automatically generated if a topic link is included in the body content of the sub-section and needs to remain.