Add action_link_content and action_link_summary to sub_sections table
We have previously renamed an existing field on the coronavirus_sub_sections table from featured_link to action_link_url.

We are now adding two additional fields to that table to complete the concept of an action link.  

These fields are: 

- action_link_content - which will be used as the link text; and 

- action_link_summary - which will be used to describe the page being linked to; the description of a content item for example.

Together, these three fields (action_link_url, action_link_content and action_link_summary) allow content designers to add action links in the accordions featured on the Coronavirus landing page where they are styled appropriately.

On the MVP UI mock-up attached to the Trello card, the action_link_summary field is shown as having a 150 character limit. The character count was queried with the designer who confirmed that the limit is not required for the MVP, but was uncertain about its need - and any limit size - for the production version.

As it is tricky to convert from a text to a string type, without loosing data - but the reverse is straightforward and lossless, and as there is uncertainty regarding this field - it is felt the best approach at the MVP stage is to use a string type for this field.