Migration to rename coronavirus_announcements path to url
This field is being repurposed to cater for URLs as well as paths. Since
paths are also valid URLs a name of url can cover both.

This migration risks breaking the application at the point of this
change (since the path field will no longer be available) until the new
code is running. Since this application has low activity I've decided to
just carefully deploy it when there aren't users rather than do a
complex backfill process for the column change.