Iterate coronavirus_announcements schema to suit data
This changes the type of the url field to be text so that it can hold
URL's longer than 255 characters, since that seems rather restrictive
now we expect absolute URLs.

It also changes a bunch of columns that should always be not null to
have the constraint. I've checked the live database and there are no
instances where these have null data.

There's a couple of improvements I didn't do as they likely require a
little more thinking, the published_at field should be a date field
rather than a datetime, and position shouldn't have a null value.