Validator for absolute paths or HTTPS urls
This validator has been added as a somewhat common case for URL's added
by GOV.UK publishers. This common case is when a URL is an absolute path
for internal URLs (on www.gov.uk) or an absolute URL to link to other
websites.

I felt the link should be a web link, which requires a scheme of HTTP or
HTTPS. I didn't think there would be good reasons why a publisher would
need link to HTTP sites so I've set this to allow HTTPS only. We can
always revise this if it becomes problematic.

The first case where this is expected to be used is in the
Coronavirus::Announcement model.