Apply AbsolutePathOrHttpsUrlValidator to Coronavirus::Announcement
This validator is applied only when the URL is not blank to avoid the
situation where you receive 2 validation messages when the field is
blank.

This also refactors some of the unit tests to validate models to use the
shoulda gem to be more succinct. I've only added very basic tests for
the AbsolutePathOrHttpsUrlValidator as this validator is already unit
tested.