Add action link fields to sub section JSON presenter
We have added attributes to the SubSection model that create the concept of an action link (a URL, a label, and a description of the page that the URL links to).  We have also added validations for this data and made it available for creation and updating by a content designer.

The data created needs to be available in the publishing-api system, so we now need to update the existing JSON payload to reflect the new data.

This change builds the JSON for an action link - if one is supplied by the user - and prepends it to the existing sub-section output in the JSON presenter.

It also adds missing tests for the sub_sections method.

The 'description_for_featured_link' and 'content_item' methods are no longer used or needed as they have been superseded by 'build_action_link'.

The 'append_priority_taxon_query_param' now takes a URL. The method is also more functional in nature as it no longer mutates a single value within a hash, but takes and returns a single value.
This method has been made private as it is not used outside of this class and is being implicitly
tested in the sub_section tests.