Refactor Announcement#published_at to make field optional
This changes the technique used to set the published_at field for
Coronavirus::Announcement so that the field can be optional and can have
improved error handling.

The previous implementation relied on setting the date to nil when it
was invalid and then providing the user with a validation error which
was caused from the date being invalid. By making this field optional a
flaw was exposed where an invalid date would produce a valid form
submission and allow the record to be saved with a date of nil (likely
to the publishers surprise).

To counter this I've moved the validation of the hash that sets the date
inside the model so that it can be a part of the other validation for
the record. This isn't the cleanest thing I've ever done by a long shot,
but since this code already relied on ActiveRecord validations it felt
like any other approach would be rather inconsistent.

I've also updated the validation rules for the date. I couldn't see any
scenario where someone would actively want a future date so I've
validated against that. I also changed the minimum time from 1950 to
2000 as it was easier to write a sensical validation message for. I
considered whether this validation was needed, given even 2000 is an
absurd date, but felt that it mostly added value in protecting against
mistakes like the year 221.