Strip the https://www.gov.uk from any announcement URLs
I chatted to Nikin about what the best approach would be to handle
publishers potentially entering https://www.gov.uk prefixed URLs into
announcements.

Our conclusion was that the friendliest approach for publishers would be
to accept the URLs rather than raise a validation error. When it came to
sending these values to the Publishing API we then strip the
https://www.gov.uk from the output so that the links are treated as
relative.