Label and hint updates for Coronavirus announcements form
I chatted with Helen and Nikin to decide what the copy changes would be
for this form.

The labels have been updated with the suggestions Hannah made [1] and an
optional suffix has been added to the publication date field to make it
clear this field is not required. This pattern doesn't seem rock solid
as I imagine it's not consistent that all non-required fields are
flagged with this, but that's what was wanted for this field.

For the hints there was a desire to specify an unenforced pattern
that brackets should be used to describe links to external sites.
The URL field now provides an external link example to indicate external
links are allowed.