Replace published_at datetime with published_on date
The published_at field for a Coronavirus::Announcement is only used to
store dates however it is configured with a datetime field.

This ports the field to be a date type and changes the name to match the
Rails convention of using a _on suffix for dates.

This includes code to backport the existing values to the new type.
Initially I just swapped the column type and renamed it, however I
discovered that this caused any dates at midnight BST to be set as the
previous day (since in the DB they are stored at UTC, so 11pm the
previous night).