Add action link URL, content and summary to the draft discarder
This is the third - and last - in a series of commits that updates the call chain that returns the current live content from publishing-api in an appropriate data structure for display back to the user on the view.

This change adds the action_link_url, action_link_content and action_link_summary fields to each sub section that is created and returned by the draft discarder.