Set position on Coronavirus::Announcement before create
This change has been made to reflect that this field no longer allows
null, instead of setting this value after the record has been persisted
to the DB we can set it before.

This approach is vulnerable to any concurrent creation of Announcements,
but then the previous implementation also was.