Stop using the github_data method to gather timeline information
The github_data method makes calls to GitHub to gather information held in yaml files. This was done so that non developers could update content. The output of this method was being modified quite opaquely to include data from timelines that is being held in the database rather than the yaml. This changes things so that the `timeline_data` method is called directly which makes it clearer that timeline data is being accessed from the database rather than the yaml.