Remove page to update the live stream
This removes the controllers, views and routes that allow users to
navigate to and update the live stream via the publishing tool.

The current live stream of press conferences used to be streamed on the
coronavirus landing page. However this stopped in summer 2020 and hasn't
been used since). The code is being removed as it's no being used and
its presence makes it harder to add features in collections-publisher.