Remove livestream editor validation helper
This helper is no longer being used now that the livestream controller
has been removed.

The current live stream of press conferences used to be streamed on the
coronavirus landing page. However this stopped in summer 2020 and hasn't
been used since). The code is being removed as it's no being used and
its presence makes it harder to add features in collections-publisher.