Remove livestream constraint on specialist sectors
There is an odd constraint that stops people with livestream editor
permissions from seeing the specialist sectors menu item. I can't think
of a reason for this except for saving space in menu, so this constraint
has been removed.

The current live stream of press conferences used to be streamed on the
coronavirus landing page. However this stopped in summer 2020 and hasn't
been used since). The code is being removed as it's no being used and
its presence makes it harder to add features in collections-publisher.