Remove live stream from fixture files
The live_stream data is no longer being used in the tests so it can be
removed from the fixture files.

The current live stream of press conferences used to be streamed on the
coronavirus landing page. However this stopped in summer 2020 and hasn't
been used since). The code is being removed as it's no being used and
its presence makes it harder to add features in collections-publisher.