Remove live stream model and factory
As we are no longer allowing users to add a live stream section to the
landing page, and are no longer sending live stream data to
publishing-api, we no longer require any of the functionality, for
example validations, that having a model provides.

The underlying database table will be used in the next commit.

The current live stream of press conferences used to be streamed on the
coronavirus landing page. However this stopped in summer 2020 and hasn't
been used since). The code is being removed as it's no being used and
its presence makes it harder to add features in collections-publisher.