Remove livestreams table from the database
No that users can no longer add a live stream section, we no longer need
this table or the data in it.

The needs to be dropped with a block to make the action reversible and
avoid the following error:

```
Rails/ReversibleMigration: drop_table(without block) is not reversible.
(https://rails.rubystyle.guide#reversible-migration, https://api.rubyonrails.org/classes/ActiveRecord/Migration/CommandRecorder.html)
```

The current live stream of press conferences used to be streamed on the
coronavirus landing page. However this stopped in summer 2020 and hasn't
been used since). The code is being removed as it's no being used and
its presence makes it harder to add features in collections-publisher.