Make SubSectionProcessor methods private
The SubSectionProcessor is used by the draft discarding service. It is
given the "sub_section" part of the payload from publishing-api, which
it uses to reverse-engineer the data needed to rebuild a sub-section in
the database.

As SubSectionProcessor has a very narrow use, it only has one caller.
This means that only the `output` method is called from outside of the
instance. Therefore it and the initializer are the only methods that
should be public. The `call` method is the only method that is being
tested as none of the tests failed when the other methods were made
private.