Make methods in `SubSectionJsonPresenter` private
Only the `output` method is called outside of the instance, so this is
the only method that should be public.

The tests have been updated to always call and test the result of the
`output` method.

To make it easier to see how the test expectations have been modified
the  structure of the tests has not been changed in this commit. That
will occur in the following commits as we try to make the tests more
maintainable.

The tests for "build link" have been removed as they aren't testing
anything new.