Stop flattening the input
There doesn't seem to be any benefit to doing this. `sub_sections`
should always be passed in as an array so, this conversion to an array
and then flattening shouldn't be needed.

The tests should be accurately representing the format of the data
passed into this class, and they are still passing.