Use inline_gem to import SVG files
The motivation for this change is the Rails deprecation warnings of:

DEPRECATION WARNING: Rendering actions with '.' in the name is
deprecated.

We previously had this issue with content-publisher. https://github.com/alphagov/content-publisher/commit/797d87106862e9a1245e63ac8a2ce75ba515a378

It  replaces the approach of using views for inline SVG files. Instead
we use the inline_svg gem which allows the SVG files to be placed in the
assets directory and then imported into views at run time. Doing this
allows us to abstract the HTML concerns (such as class and aria-hidden)
from the SVG and handle it in the view.

The additional config  will cause an exception to be raised if an Inline SVG
is embedded and the file doesn't exist.