Fix flaky curating topics test
The test is failing with the following error:

```
Failures:

 ​1) Curating topic contents Curating the content for a topic with javascript
    ​Failure/Error:
      ​expect(titles).to eq([
        ​"Oil rig staffing",
        ​"Oil rig safety requirements",
      ​])

      ​expected: ["Oil rig staffing", "Oil rig safety requirements"]
           ​got: ["Oil rig safety requirements", "Oil rig staffing"]

      ​(compared using ==)
    ​# ./spec/features/curating_topic_contents_spec.rb:73:in `block (4 levels) in <top (required)>'
    ​# /var/lib/jenkins/bundles/ruby/2.7.0/gems/capybara-3.35.3/lib/capybara/session.rb:353:in `within'
    ​# /var/lib/jenkins/bundles/ruby/2.7.0/gems/capybara-3.35.3/lib/capybara/dsl.rb:53:in `call'
    ​# /var/lib/jenkins/bundles/ruby/2.7.0/gems/capybara-3.35.3/lib/capybara/dsl.rb:53:in `within_element'
    ​# /var/lib/jenkins/bundles/ruby/2.7.0/gems/capybara-3.35.3/lib/capybara/rspec/matcher_proxies.rb:15:in `within'
    ​# ./spec/features/curating_topic_contents_spec.rb:68:in `block (3 levels) in <top (required)>'

Finished in 38.94 seconds (files took 4.69 seconds to load)
993 examples, 1 failure
```

The test is checking that if a user has javascript and goes to one of
the [specialist sector pages][1] that they can drag items from the left
hand side into a list on the right.

Because the `drag_to` function is looking for a `tbody` each item should
be dragged to the top of the list.

Therefore if, "Oil rig staffing" is dragged first and "Oil rig safety
requirements" second, the test expects that they appear in this order on
the page, though I'm not sure why:

* Oil rig staffing
* Oil rig safety requirements

But this doesn't happen consistently. When the test is run locally it
passes. It only seems to fail with any consistency on CI.

There are two PRs being blocked by this failure:

* https://github.com/alphagov/collections-publisher/pull/1435
* https://github.com/alphagov/collections-publisher/pull/1434

What else was tried (and failed)?
1. The test [waits for Ajax][2] to complete before doing the next drag
and drop. The default value for `Capybara.default_max_wait_time` is 2
seconds. I've tried adding this value to `spec_helper` and setting it to
"50".

2. Changing the `eq` on line 73 to `match_array`, however this just
means the test fails later on on line 100 when the payload for
publishing-api is checked.

3. Running the test 1000 times locally. The test failed 7 times in
total, but it was the call to publishing-api that failed because the
list items were in the incorrect order:

```
7) Curating topic contents Curating the content for a topic with javascript 293
     Failure/Error:
       assert_publishing_api_put_content(
         content_id,
         request_json_includes(
           "details" => {
             "groups" => [
               { "name" => "Oil rigs",
                 "contents" => [
                   "/oil-rig-staffing",
                   "/oil-rig-safety-requirements",
                 ] },

       The request PUT https://publishing-api.test.gov.uk/v2/content/5a4df1e2-e1f6-4a6c-a960-4b99ba34f219 with given block was expected to execute 1 time but it executed 0 times
```

A `wait_for_ajax` [3] method was added because the test was breaking
intermittently due to a race condition. Whilst that is was fixed the
test continues to fail intermittently. This could be because whilst the
list items are being dragged to `tbody`, we're not specifying where in
`tbody`. As the code doesn't actually say "put x to the top of the list"
in the test name I think it's ok to use non-ordering specific matchers.

The assertion that "put content" is being called in publishing-api has
also changed as the order of the items in the payload depends upon the
order of the items in the database, not the order of the items on the
screen.

[1]: https://collections-publisher.integration.publishing.service.gov.uk/tags/a32f4ae1-5400-4807-ab4c-5e9fb57ea5dd/lists
[2]: https://github.com/alphagov/collections-publisher/blob/e434823c10952d6c1fa885593e6d4ef94407549b/spec/support/wait_for_ajax.rb#L3
[3]: https://github.com/alphagov/collections-publisher/commit/a52db3bbf5161ce17d3204296cd5ef22ad956e98