Add a task to unpublish the coronavirus hub pages
We have been given permission to retire the coronavirus hub pages.
The links to the hub pages have already been removed and the content in
the hubs as already been added to the accordions.

This task is temporary and will be removed in the story to remove all
hub-related code.