Add a task to remove the coronavirus hub pages
We have been given permission to retire the coronavirus hub pages.
The links to the hub pages have already been removed and the content in
the hubs as already been added to the accordions and the hub pages have
been unpublished, so the database entries can be removed.

This is a temporary task that will be removed in a follow-up ticket.