Remove coronavirus rake tasks
The `remove_hub_pages` is no longer necessary now that the hubs have
been removed.

Also removes the `resync_titles` task. This task looks like it was
needed whilst the publishing tool was being built to pick up any changes
to the page and accordion titles.

I can't remember the last time this task was run, so I don't think it's
needed. ContentBuilder picks up the entire contents of the GitHub YAML
file [1] and sends it publishing-api, however as a precaution
`sections_heading` has been added as a required field to the GitHub
validation so that an error will be thrown if the field is accidentally
removed from the YAML file

[1]: https://github.com/alphagov/collections-publisher/blob/278a10bb7c512cfd86eaf1a694b6cd887357d9c3/app/services/coronavirus/pages/content_builder.rb#L16