Controller actions for editing header
`edit_header` and `update_header` were chosen rather than the standard
`edit` and `update` because this would allow us to create separate edit
actions for editing other landing page elements in the future if needs
be. The path name is also more explanatory to the user.

As far as I can tell, now that the hub pages are gone, we no longer need
to find page by slug. A commit to change the `topic_page` index variable
to use `@page` will be added in a future commit.

Using locale files for the notice text will also be added in a future
commit.