Validate header link has both url and link text
Should be valid if both the header action link url and text have values,
or if neither has a  value.

Uses the XOR operator to determine if only one of the fields is
populated. This is an interesting operator, as once you've determined
that both of the fields are populated / not-populated, you only need to
check the status of one field at a time to check whether the field is
valid.

For example:

```
irb(main):001:0> a = true
=> true
irb(main):002:0> b = true
=> true
irb(main):003:0> a ^ b
=> false

irb(main):004:0> a = false
=> false
irb(main):005:0> b = false
=> false
irb(main):007:0> a ^ b
=> false

irb(main):008:0> a = true
=> true
irb(main):009:0> b = false
=> false
irb(main):010:0> a ^ b
=> true

irb(main):011:0> a = false
=> false
irb(main):012:0> b = true
=> true
irb(main):013:0> a ^ b
=> true
``