Add task to sync coronavirus header with YAML
Adds a rake task to sync the header section on the coronavirus landing
page with the entries in the YAML file.

Creating a rake task avoids the risk of copy and paste errors that could
occur if the section was updated manually.

This is a throwaway task that should only be run once and will be
removed after the header section has been deployed.

Usage:

```
bundle exec rake coronavirus:sync_header_section
```