Only present header if unreleased features is set
This prevents the header section being presented from the database if
the "unreleased features" flag has not been set in an environment.
The "unreleased features environment" variable has only been set in
integration [1] to allow for testing. Using this flag allows us to push
the code to production and only "turn it on" when we're ready for the
feature to go live.

[1]: https://github.com/alphagov/collections-publisher/pull/1192/commits/300523c17ec8ed8da4b59deff84d5e374855b9e9