Merge Browse Pages into Topics
A naive approach that creates a new Specliast Topic for every
mainstream browse page. This is only a prototype and should not be
pushed beyond this point without serious work.

Known issues: Children are created as-is without consideration of
them being a different type. Next rendition will likely have it fixed
with iteration  rather than blanket assignment, however discussion with
content means this may be a dead end...