Update edit link text
Change edit link label text from "Change" to "Edit".

The word "Change" causes the action column to wrap around due to the component not being responsive. Previously we had css overrides but this was removed by PR #1539.

While we wait for the component to update layout, this change will help keep a nice consistent layout.