Disentangle where the different layouts intersect
This application runs with two different layouts: a govuk_admin_template
(legacy) one and a govuk_publishing_components/design system one.

The previous configuration for the design system one had a mixture of JS
and SCSS from the legacy and the current. This commit fixes that.

The things this commit does is:

- Rename the layout files to distinguish between design system and legacy
- Update calls to layout so that we just specify when we want the legacy
  layout (since the majority is design system)
- Renamed the JS and CSS files because the layout_for_admin design
  system component expects an application.js and application.css files
  and this is not configurable
- Rename the legacy CSS and JS files to be application_legacy to reflect
  the legacy status.