Add tag and title helper methods for the GOV.UK Design System
This commit simplifies the approach used for generating the tag
and title helper methods that were implemented in bootstraps.

While it isn't a like for like copy it seems like all pages should
show a tag, rather than conditionally like they are at the moment.
Happy to be challenged on this!

I've added the ability to append or prepend a title.

Also, we handle breadcrumbs in a separate component.