Stop adding GitHub content to presenter payload
Removes GitHub content from ContentBuilder. ContentBuilder is used to
collate the content from the GitHub YAML file and the
collections-publisher database. PagePresenter, which was updated in the
previous commit, then creates the payload from publishing-api from this
data. No that the content is being hardcoded in the frontend, we no
longer have to collate this content.

"title" is still a required field in the content item schema for the
coronavirus landing page, even if it's going to be hardcoded into the
frontend, so we must pass it in the presenter. If we don't pass the
title the schema validation test in page_presenter_spec against
govuk-content-schemas fails.