Remove GitHub content from PagePresenter
The title and description are populated from the GitHub YAML file in
govuk-coronavirus-content repo. However, as this repo is being archived,
and the content to be hardcoded in the collections app, we need to
decide what to do about the title and description.

The description is an optional field in the content item so it has been
removed. However, the title is mandatory and will need to still be
populated even if the value is not used in the frontend, otherwise the
schema checks in publishing-api will fail. The title will be hardcoded
into ContentBuilder in the next commit. This commits prepares for that
change by pulling in the title content from ContentBuilder in the tests.