Remove the GitHub changes pages
Removes the Github changes controller, views and routes. This is the
page that allowed content designers to pull in the YAML changes from
govuk-coronavirus-content repo and preview the changes before
publishing. This content is going to be hard-coded in the collections,
so this page is no longer needed.

Also removes the locale content and feature tests.