Add mainstream_browse_origin to TopicPresenter
We want to reference the content_id of Mainstream browse page the topic
originated from as a part of unified topic system work.

Until the work is completed we will need to keep the Mainstream browse page and
the cloned Topic is sync. Since other attributes of the Mainstream browse page
such as title or slug can change, this adds a unique reference we can use to
identify the cloned Topic.

This attribute will be populated when a new Topic is created and send to the
Publishing API.
The presence of this attribute will be used to hide those Topic.

Following the content schema change in: https://github.com/alphagov/govuk-content-schemas/pull/1089