Copy curated lists into the new Topic
Curated lists / groups are published separately to the Topics itself.
The topic needs to be published before the ListPublisher is called
as it fetches the tagged_documents utilising Services.publishing_api.get_linked_items
If we don't publish the Topic first we'll get 404 No item with content_id of the new Topic.