Add the list show page
This adds a show page for a list. It displays all the list items
ordered by their indexes. Each list item links to the corresponding
frontend page, and has links for moving it to another list and deleting
it.

The latter two bits of functionality have not been wired in yet.