Add `available_list_items` method to `List`
A requirment of adding additional linked list items to a list is the
ability to filter out list items that already are present on the list.

This method adds this functionality by calling the publishing api for
tagged documents, then filtering out any present on the list by removing
any with the same base path as a present list item.