Add functionality for adding items to a list
This commit adds the `edit_list_items` endpoint. It renders a
list of checkboxes of available list items that can be added to a
list.

It does this by filtering out list items that are currently in the list
from the a topic or browse pages tagged documents based on a list
items `base_path`.