Remove redundant code that dirties tags
At the moment, each list controller action marks a tag as dirty,
then publishes the list. Publishing the list then sets dirty to false.

This is a hangover from when everything was drag and drop one one page
and lists were updated dynamically without forcing the user to publish
the result.