Update error items helper to handle checkbox component
The error handling for this component is built slightly differently
to the other components. Resultantly, an empty array is rendered
as an error message for the checkbox component. This is due to there
not being a presence check within the component before errors are
passed back.

This is easily mitigated on our end by ensuring that objects passed into
the errors_for method return nil when errors are not present.