Ensure untagged objects links fail gracefully
At the moment, if someone untags a document on content tagger it causes
an error to occur when you attempt to link to the document on content
tagger on the list show page. This happens as the link to content tagger
is located by parsing the tags tagged documents for a content id to
construct the path.

This updates the link so that if a content id cannot be found it passes
them to the lookup page on Content Tagger where they can locate the
link and update the tagging.