Find subscriber list slug for topic before unsubscribing
Specialist topic slug does not always match Subscriber list slug for this
topic.  E.g. for /topic/benefits-credits/child-benefit the slug in Collections
Publisher is child-benefit but Subscriber list slug is
tax-credits-and-child-benefit-child-benefitand and there is no content_id
associated with this Subscriber list.

For this reason we need to make an additional call to find the subscriber list
by links, which uses
https://github.com/alphagov/gds-api-adapters/blob/58bd9a8/lib/gds_api/email_alert_api.rb#L22-L28

The EmailAlertsUnsubscriber can be easily used for archiving Coronavirus pages.